refactor: posts를 affine으로 변경
* 기존 posts는 supabase에 markdown을 올리고 로드하는 방식임
* markdown을 작성하고, exceprt 등의 형식을 맞추는 것이 너무 번거롭고 글을 안 쓰게됨
* 테스트해보니 괜찮은거 같아서 바꿈
