+++ app/posts/page.tsx
@@ -1,60 +1,43 @@
-import { createClient } from "@/utils/supabase/server";
 import { Box, Text } from "@mantine/core";
-import dayjs from "dayjs";
+import { getBlocksuiteReader } from "affine-reader";
 import Link from "next/link";
+import dayjs from "dayjs";
 
-interface Post {
-  title: string;
-  excerpt: string;
-  date: string;
-}
-
-async function getPosts() {
-  const supabase = createClient();
-
-  const { data, error } = await supabase
-    .from("blog")
-    .select("*")
-    .order("created_at", { ascending: false })
-    .range(0, 9);
-
-  if (error != null) {
-    console.log(error);
-    return [];
-  }
-
-  const promises = data.map(
-    async (item): Promise<{ file: string; frontmatter: Post }> => ({
-      file: item.file,
-      frontmatter: {
-        title: item.title,
-        excerpt: item.description,
-        date: item.created_at,
-      },
-    })
-  );
-
-  const list = await Promise.all(promises);
-
-  return list;
-}
-
-export default async function Posts() {
-  const posts = await getPosts();
+export default async function Affines() {
+  const reader = getBlocksuiteReader({
+    workspaceId: "834008fb-68de-4274-9f69-dbe8ae03d274",
+  });
+  const metadata = await reader.getDocPageMetas();
 
   return (
     <>
-      {posts.map(({ file, frontmatter: { title, excerpt, date } }) => (
-        <Box key={title} maw={840} my="md">
-          <Text component={Link} href={`/posts/${file}`} fw={700} fz="xl">
-            {title}
-          </Text>
-          <Text c="dimmed">{excerpt}</Text>
-          <Text c="dimmed" fz="xs">
-            {dayjs(date).format("YYYY-MM-DD")}
-          </Text>
-        </Box>
-      ))}
+      {metadata
+        ?.filter((p) => !p.trash)
+        ?.sort((a, b) => {
+          const dateA =
+            a.properties?.custom.find((c) => c.name === "created")?.value ??
+            a.createDate;
+          const dateB =
+            b.properties?.custom.find((c) => c.name === "created")?.value ??
+            b.createDate;
+          return dayjs(dateB).valueOf() - dayjs(dateA).valueOf();
+        })
+        .map(({ title, id, createDate, properties }) => (
+          <Box key={title} maw={840} my="md">
+            <Text component={Link} href={`/posts/${id}`} fw={700} fz="xl">
+              {title}
+            </Text>
+            <Text c="dimmed">
+              {properties?.custom.find((c) => c.name === "excerpt")?.value}
+            </Text>
+            <Text c="dimmed" fz="xs">
+              {dayjs(
+                properties?.custom.find((c) => c.name === "created")?.value ??
+                  createDate
+              ).format("YYYY-MM-DD")}
+            </Text>
+          </Box>
+        ))}
     </>
   );
 }
